backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
scikit-learn==1.7.2
pandas==2.3.2
numpy==2.2.4
joblib==1.5.2
imbalanced-learn==0.14.0
pymongo==4.10.1
python-dotenv==1.0.0
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0